pyproject.toml
[tool.poetry]
name = "tenderpost-backend"
version = "1.0.0"
description = "Production-grade tender scraper using Crawl4AI + Playwright + 2Captcha"
authors = ["TenderPost Team"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
crawl4ai = "^0.7.0"
playwright = "^1.40.0"
httpx = "^0.27.2"
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
pillow = "^10.1.0"
selectolax = "^0.3.17"
python-dotenv = "^1.0.0"
supabase = "2.24.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
black = "^23.11.0"
mypy = "^1.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py39']

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

